=htc extract HttpServerTestSetupBase from HttpServerSpec
